Move util shared by client and adapters out of adapters subpackage.
